# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Node-Typescript API",
    "version": "1.0.0",
    "description": "A sample API"
  },
  "servers": [
    {
      "url": "http://192.168.0.119:3000"
    },
    {
      "url": "http://localhost:3000/"
    },
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "http://ec2-18-216-130-99.us-east-2.compute.amazonaws.com:3000"
    },
    {
      "url": "http://ec2-3-22-60-204.us-east-2.compute.amazonaws.com:3000"
    }
  ],
  "paths": {
...
    "/api/customer/{id}": {
      "get": {
        "description": "get submaterials for the given materialId",
        "tags": [
          "customer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": {
                    "$ref": "#/components/schemas/CustomerSchema"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "delete customer by id",
        "tags": [
          "customer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "content": null
          }
        }
      }
    },
...
    "/api/project": {
      "get": {
        "description": "Get Projects api with pagination and sorting",
        "tags": [
          "project"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Example = active"
          },
          {
            "in": "query",
            "name": "projectName",
            "schema": {
              "type": "string"
            },
            "description": "Example = project 1"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": 1
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": 10
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "Example = createdAt"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "acs"
            },
            "description": "Example = acs"
          }
        ],
        "responses": {
          "200": {
            "description": "returns projects list",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProjectSchema"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "create project",
        "tags": [
          "project"
        ],
        "requestBody": {
          "description": "project creation request body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectSchema"
              },
              "example": {
                "name": "userName",
                "description": "desc",
                "customerId": "5f9cdeaf63bcde326cc86b1e",
                "location": {
                  "city": "oslo",
                  "postCode": 548
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "return created project",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProjectSchema"
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "update project",
        "tags": [
          "project"
        ],
        "requestBody": {
          "description": "_id field is required",
          "required": true,
          "content": {
            "application/json": {
              "required": [
                "id"
              ],
              "schema": {
                "$ref": "#/components/schemas/ProjectSchema"
              },
              "example": {
                "_id": "5f9bbb17405a9420e895920f",
                "contactPersonName": "Mohcine",
                "location": {
                  "_id": "5f9d5344157c674bf0798c93",
                  "city": "somelo",
                  "postCode": 548
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "return created project",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProjectSchema"
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/{id}": {
      "get": {
        "description": "get submaterials for the given materialId",
        "tags": [
          "project"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": {
                    "$ref": "#/components/schemas/ProjectSchema"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "get submaterials for the given materialId",
        "tags": [
          "project"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "content": null
          }
        }
      }
    },
    "/api/project/customer/{id}": {
      "get": {
        "description": "get project by customer id",
        "tags": [
          "project"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": {
                    "$ref": "#/components/schemas/ProjectSchema"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CustomerSchema": {
        "properties": {
          "userId": {
            "type": "string"
          },
          "contactPerson": {
            "type": "string"
          },
          "customerType": {
            "type": "string"
          },
          "contactPersonNumber": {
            "type": "string"
          },
          "organizationNumber": {
            "type": "string"
          },
          "contactPersonEmail": {
            "type": "string"
          },
          "locationId": {
            "type": "string"
          },
          "location": {
            "type": "object",
            "$ref": "#/components/schemas/LocationSchema"
          },
          "user": {
            "$ref": "#/components/schemas/UserSchema"
          }
        }
      },
...
      "LocationSchema": {
        "properties": {
          "name": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "stateCode": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "postCode": {
            "type": "string"
          },
          "formattedAddress": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          }
        }
      },
...
      "ProjectSchema": {
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "contactPersonName": {
            "type": "string"
          },
          "contactPersonEmail": {
            "type": "string"
          },
          "contactPersonPhoneNumber": {
            "type": "string"
          },
          "projectNumber": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "estimatedHours": {
            "type": "number"
          },
          "registeredHours": {
            "type": "number"
          },
          "startDate": {
            "type": "date"
          },
          "completionDate": {
            "type": "date"
          },
          "deadLine": {
            "type": "date"
          },
          "locationId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "location": {
            "type": "object",
            "$ref": "#/components/schemas/LocationSchema"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerSchema"
          }
        }
      },
...
      "UserSchema": {
        "required": [
          "email",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "passwordResetToken": {
            "type": "string"
          },
          "passwordResetExpires": {
            "type": "string",
            "format": "date"
          },
          "tokens": {
            "type": "array"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "Error description",
            "example": "User created"
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-access-token"
      }
    }
  },
...
}